Moved the createUser and the login controllers to the authController.js
